Refining and restructuring the test cases
