+++ Tests/Schema.NET.Test/Examples/OrganizationTest.cs
@@ -1,7 +1,6 @@
-namespace Schema.NET.Test
+namespace Schema.NET.Test.Examples
 {
     using System;
-    using Newtonsoft.Json;
     using Xunit;
 
     // https://developers.google.com/search/docs/data-types/corporate-contacts
@@ -38,22 +37,6 @@ namespace Schema.NET.Test
             "\"logo\":\"https://example.com/logo.png\"" +
         "}";
 
-        private readonly string jsonHttp =
-        @"{" +
-            "\"@context\":\"https://schema.org\"," +
-            "\"@type\":\"Organization\"," +
-            "\"url\":\"https://example.com\"," +
-            "\"areaServed\":\"GB\"," +
-            "\"contactPoint\":{" +
-                "\"@type\":\"ContactPoint\"," +
-                "\"availableLanguage\":\"English\"," +
-                "\"contactOption\":\"https://schema.org/TollFree\"," +
-                "\"contactType\":\"customer service\"," +
-                "\"telephone\":\"+1-401-555-1212\"" +
-            "}," +
-            "\"logo\":\"https://example.com/logo.png\"" +
-        "}";
-
         [Fact]
         public void ToString_CorporateContactsGoogleStructuredData_ReturnsExpectedJsonLd() =>
             Assert.Equal(this.json, this.organization.ToString());
@@ -61,14 +44,8 @@ namespace Schema.NET.Test
         [Fact]
         public void Deserializing_OrganizationJsonLd_ReturnsOrganization()
         {
-            Assert.Equal(this.organization.ToString(), JsonConvert.DeserializeObject<Organization>(this.json, TestDefaults.DefaultJsonSerializerSettings)!.ToString());
-            Assert.Equal(JsonConvert.SerializeObject(this.organization, TestDefaults.DefaultJsonSerializerSettings), JsonConvert.SerializeObject(JsonConvert.DeserializeObject<Organization>(this.json, TestDefaults.DefaultJsonSerializerSettings), TestDefaults.DefaultJsonSerializerSettings));
+            Assert.Equal(this.organization.ToString(), SchemaSerializer.DeserializeObject<Organization>(this.json)!.ToString());
+            Assert.Equal(SchemaSerializer.SerializeObject(this.organization), SchemaSerializer.SerializeObject(SchemaSerializer.DeserializeObject<Organization>(this.json)!));
         }
-
-        [Fact]
-        public void Deserializing_OrganizationJsonLd_WithHttpSchemaOrg_ReturnsOrganization() =>
-            Assert.Equal(
-                this.organization.ToString().Replace("\"@context\":\"https://schema.org\"", "\"@context\":\"https://schema.org\"", StringComparison.Ordinal),
-                JsonConvert.DeserializeObject<Organization>(this.jsonHttp, TestDefaults.DefaultJsonSerializerSettings)!.ToString());
     }
 }
