--- Tests/Schema.NET.Test/core/TestDefaults.cs
@@ -1,19 +1,0 @@
-namespace Schema.NET.Test
-{
-    using Newtonsoft.Json;
-    using Newtonsoft.Json.Serialization;
-
-    public static class TestDefaults
-    {
-        public static readonly JsonSerializerSettings DefaultJsonSerializerSettings = new()
-        {
-            Formatting = Formatting.None,
-            DateParseHandling = DateParseHandling.DateTimeOffset,
-            DefaultValueHandling = DefaultValueHandling.Ignore,
-            NullValueHandling = NullValueHandling.Ignore,
-            TypeNameHandling = TypeNameHandling.None,
-            ContractResolver = new CamelCasePropertyNamesContractResolver(),
-            ReferenceLoopHandling = ReferenceLoopHandling.Ignore,
-        };
-    }
-}
